CP Algorithms: global comparison of block configuration and sequences
This is hopefully very straightforward, but also very incomplete: The
idea is that to validate the config blocks, I produce the same n-tuple
once with the config blocks and once with the sequences and then
compare it.  That validates that (at least for the chosen settings and
variables) both configurations are the same.  As a bonus, once
completed this also runs a fairly complete CP algorithms job in CI.

So far this is very incomplete, it only runs muons and only checks the
four-momentum, but having this in the release should allow adding more
objects and variables as I add more configuration blocks.