HltEventLoopMgr: add option to run algs/sequences during start
Add the `execAtStart` property that can be used to specify a list of
algorithms/sequences that should be executed at the end of the
prepareForRun transition. This can e.g. be used to execute all CondAlgs
once before the first event to improve memory sharing:

ServiceMgr.HltEventLoopMgr.execAtStart = ['AthCondSeq']

Relates to ATR-26138.