LAr SC HV corr computation
This should add possibility, to compute the HV corrections for the LAr digital trigger. Also some cleanup and fixes for code used to compute, store and dump HV corrections for both main readout and digital trigger one. Code is not used in the standard workflows, only in the conditions manipulations jobs.