Updating some derivation ARTs in master
Updating some derivation ARTs

This MR updates the preExec of some ART which are still missing the additional flag 'DetFlags.detdescr.Muon_setOn();'. Some train tests have beem moved to the right folder (not sure why the ended up in the EGAM one). I took this opportunity to activate the sharedWriter tests in master/AthDerivation (currently they are only running in master/Athena). master/AthDerivation seems to be running fine (except for a couple of failures we need to follow up), so it may be fully operational soon (after that master/Athena will all be disabled).