Consistently use CommonAugmentation instead of DerivationKernel for common DAOD augmentations
Consistently use CommonAugmentation instead of DerivationKernel for common DAOD augmentations

We noticed during the CA migration of the derivation framework that the cut flow book-keepers were registering cuts from a variety of algorithms that were not involved in event selection. This was because in a few places DerivationKernel (an AthFilterAlgorithm with setFilterPassed explicitly set) was used in writing common augmentations, whereas the correct thing to do is to use CommonAugmentation, which does the same job but doesn't register a cut. This MR implements that correction. Although this issue impacts both the legacy configuration and the CA version, the fix is only applied in the CA.