Fixing warnings in derivation jobs
Fixing warnings in derivation jobs

The first warnings in InDetJetFitterUtils are normal things that happen
regularly in good data and MC events. The warnings were downgraded to
DEBUG messages in !27971 , and this set of changes basically brings
those forward to master.

The second warnings in AutoConfiguration build on those changes in
!29679 , reducing to info messages a number of things that are perfectly
normal during running with modern releases and modern input files.