Bmumux triggers: Defined Zed Width of RoI and set default value for Bmumux triggers
Bmumux: Defined Zed half-width and set default value for Bmumux triggers

Defined the z half-width for Bmumux triggers in BphysicsMenuSequence.py
so that it can be optimised.
Set the z half-width to the current default value of 225 within ConfigSettings.py
extending along the full interaction region of the beam pipe.