Updating derivation ARTs in master
Updating derivation ART in master

This MR updates the ARTs in master/Athena and master/AthDerivation with the most updated preExec and postExec. It also add a new test (EGAM10). I took this opportunity to reorganise a little bit the code structure to be slightly more readable (i.e. I split the Reco_tf command parameters in multiple lines).