LArG4Code: do not delete ILArCalculatorSvc
LArG4SimpleSD and LArG4CalibSD are deleting the `ILArCalculatorSvc` on
their destruction claiming they are "owning" it. That is certainly
wrong as services are not owned by an individual component. Likely
didn't cause problems so far because the G4 RunManager who ultimately
triggers this destructor is not being cleaned up properly either.