introduce (optional) per-algorithm timers for EventLoop
The main goal of these is to find hot spots in the CP algorithm
sequences that can then be debugged with other tools.  These timers
introduce almost no overhead and it's fairly straightforward to add
them to full-length jobs to see where time is spend.

For the reporting there is just an individual print-out per algorithm
at the end of the job.  Some nicer output could be added in the future
if this gets used on a more regular basis.