add "proper" views and thinning for full CP algorithm test sequence
This removes all the implicit views created by the various sequences,
and instead creates explicit views as inputs for MET and OR.  And it
adds actual consistent thinning for all containers for the output.
And it adds pt/eta cuts for all the objects, which are probably way
off from what they should be...

I also had to add some extra configuration options/decorations to some
of the sequences to make it all consistent.