xAOD::TEvent Speedup, master branch (2022.08.26.)
Made it faster for analysis code to repeatedly check for missing objects.

When analysis code uses xAOD::TEvent::contains<...>(...) to check if it
should do something, the old code ended up calling TTree::GetBranch(...)
for missing containers/branches, every time that xAOD::TEvent::contains
was called.

First off the TTree::GetBranch(...) call was replaced by a faster search
among the top-level branches of the tree.

But even more importantly, xAOD::TEvent was taught to remember what
branches/objects it already looked for, but didn't find. So that on
subsequent calls it would be able to determine more quickly that the
object/branch is (still) missing.