BS vs Mu plot updates for Run 3
BS vs Mu plot updates for Run 3

Tidying and a few updates to the BS vs Mu plot maker for Run 3. Mostly
this is a pile of style changes (for consistency; spacing before was
inconsistent). Also fixing a few typos (mostly in comments or help doc),
and adding information for running on Run 3 data.