AthMessaging: fix for lazy MessageSvc initialization
The `setLevel()` method triggered the MessageSvc retrieval and if not
available prints a WARNING message. This is because we were re-using the
message level `NIL` as an initialization flag. Add an explicit flag for
this use-case so `setLevel()` becomes a trivial method again.

Also re-order member variables to avoid padding.