Fixing a bug which causes some containers to be omitted from CA-produced DAODs
Fixing a bug which causes some containers to be omitted from DAODs

This MR fixes a bug in `TrigNavSlimmingMTConfig.py` which, whilst trying to add trigger-related containers to the slimming dictionary, in fact overwrites it. The slimming dictionary is used to map interface and aux containers and their types, and only containers appearing in it will be written to the DAODs, so in princple all of the overwritten containers would be dropped from the DAOD. However, many of the overwritten containers were then restored via the ContainersOnTheFly content, but anything not in this list was finally dropped.

The MR makes two changes:
- fixes the dictionary over-write by using the `update` method
- adds several truth containers, which are now part of the common physics content, to the `ContainersOnTheFly.py` list, meaning they don't have to be explicitly added in the format definition

With this change in place the expected containers are restored.