Make reco integration tests lighter to avoid timeouts
tauRecTools: drop conditional decoration in TauIDVarCalculator

Hello,

This is resubmission of the broken !56154 .

We have a conditional decoration in tauRecTools, i.e. a decoration that is only added onto certain taus (those with at least 1 track), which causes trouble downstream when making derivations.
This decoration corresponds to the eProbabilityHT of the leading track of the tau.
It was created in R21 for convenience, but it is no longer used anywhere in R22 (and it can anyway be recomputed from the AOD).
As agreed at the last RIG meeting, this decoration is now dropped.

The ESD and AOD reference files for q442, q445 and q449 have been updated and deployed on EOS.
The AOD_content.ref files have been updated as well.

Cheers,
Bertrand