Add code to allow DAOD_PHYSLITE to be made from PHYS with the component accumulator
Draft: add code to allow DAOD_PHYSLITE to be made from PHYS with the component accumulator

This MR adds the small amount of extra python needed to allow PHYSLITE to be made from PHYS using the Derivation_tf transform, that is, the component accumulator. It touches three parts:

1. MuonAnalysisSequence has an extra argument allowing the internally produced MuonSelectionTool to be set to run 3 geometry should the input be data18/mc20. If this isn't done the tool crashes. This doesn't happen with AOD->PHYSLITE, presumably because the properties are inherited from the MuonCommonConfig, but this isn't used for PHYS->PHYSLITE
2. DerivationSkeleton is slightly adjusted to accommodate the different arguments, and also to create a new ConfigFlag called Input.FileType (currently can be AOD, DAOD_PHYS, EVNT)... this may be useful for other applications as well
3. PHYSLITE.py adjustment itself to put in guards to ensure that code that only works with AOD input (e.g. writing common physics content and re-setting MET association) isn't scheduled. It largely uses the new FileType ConfigFlag to achieve this. Also it sets the relevant geometry in the MuonAnalysisSequence.

With these changes, the following command works

```
Derivation_tf.py --CA --inputDAOD_PHYSFile filename.pool.root --outputD2AODFile output.pool.root --formats PHYSLITE
```

On account of the disabling of the MET association, the output is not yet usable for physics analysis. However, it still makes sense to prepare this MR to allow testing of this workflow, which will be needed during run 3.