TrkVertexSeedFinderUtils: Don't use exceptions in GetClosestPoints.
TrkVertexSeedFinderUtils: Don't use exceptions in GetClosestPoints.

Throwing an exception is expensive, so exceptions should be used only
for things that are really exceptional.
This wasn't the case for GetClosestPoints.
Change it so that it can optionally return an error string rather than
the regular result instead of raising an exception.