Include SimulationConfig package in AnalysisBase
Incclude SimulationConfig package in AnalysisBase

This follows up on a MR which added ConfigFlags to AnalysisBase.
Since SimulationConfig is a very light-weight package that was
made so it could be used by ConfigFlags in AthAnalysis, it likely
makes sense to add it to AnalysisBase as well. This gives users
the ability to check ConfigFlags for things like simulation flavour
in the same way as in Athena.