Patch to get_files to use mktemp
Patch to get_files to use mktemp

As suggested in ATEAM-858, patching get_files to use mktemp. Seems to be
working in a few local tests, and seems to be a well-supported command.
This avoids making use of /tmp in systems that might not have /tmp and
might not have set $TMPDIR (e.g. might have set $TEMPDIR, or might use
some other path for temporary files).