TauAnalysisTools: downgrade FATAL to WARNING in BuildTruthTaus
TauAnalysisTools: downgrade FATAL to WARNING in BuildTruthTaus

Hello,

As discussed in ATLASG-1661, when building TruthTaus from the truth particle container in Herwig7, it seems that tau decays involving an intermediate a1 meson are buggy: the a1 doesn't decay, which creates problems in BuildTruthTaus when we navigate through the tau decay products. When an unexpected decay mode is found, we call a function that recursively prints the daughther of the tau, and we encounter a second "bug/feature": some status-1 photons have a decay vertex with 2 daughters, but these daughters are not present in the truth record (there is suspicion that it may come from photon converions in Geant4). When this happens, we print a FATAL error message, but we abort the execution with a FAILURE, it's just a printout. However this causes derivation jobs to fail when parsing log files.

This MR downgrades the FATAL to a WARNING. But there seem to be an issue with Herwig7(+G4?), which will have to be followed up with experts.

Adding the urgent label as it is blocking an imminent DAOD production.

Cheers,
Bertrand