LArDigitContainer: introduce p3 persistent representation and associated TP converter
LArDigitContainer: introduce p3 persistent representation and associated TP converter

Hello,

As discussed in ATR-26276 and ATR-25881, this MR is introducing a new version for the LArDigitContainer persistent representation (p3 which supersedes p2), to handle both standard LAr cells and supercells.
The name of the transient container is used to determine whether it's standard or supercells (in the latter case, the container name contains 'SC' by convention).
The p3 converter design is completely based on p2, except that we now switch between LArOnlineID and LArOnline_SuperCellID helpers to do the mapping between IdentifierHash and HWIdentifier in the transient/persistent conversions, depending whether it's standard cells or supercells.

Cheers,
Bertrand