Improve start parameters of TRT segment track fit.
Improve start parameters of TRT segment track fit.

Improve start parameters of TRT segment track fits for tracks predominantly
traversing the TRT endcaps which have an initial theta estimate close to zero,
which may happen in rare cases.

The computation of the start parameters requires an estimate of theta which is
now computed from the radial position of the first and last endcap measurements.
If this still results in a theta close to zero, theta is estimated from the
maximal radial distance between the inner or outermost radial extent of the first
and last endcap TRT straw on track.