TrigDecisionTool: add isPassed[Bits]ForEach to ChainGroup
Add `isPassedForEach` and `isPassedBitsForEach` function to `ChainGroup`
to allow the retrieval of the trigger decision (bits) of each chain in the group.
Also optimize the `isPassed` function to return at the first passing chain
instead of continuing to loop over all chains.

In order to maintain the correct order between trigger names and the
result of the "ForEach" methods, the storage of the chains/items had to
be changed from `std::set` to `std::vector` (not clear why a set was
used in the first place).

Closes ATR-26331.