TestHepMC Patches and Warning cleanup
TestHepMC Patches and Warning cleanup

Adding back a patch to master that was in !36323 (thanks to Jason Veach
for the report).

While I'm here, tidying up a bunch of warning messages that were
bothering me a bit (things that printed blank lines, things that printed
a header but no actual output, etc).