LArDigitContainerCnv: replace LArDigitContainer_p2 by p3 as new default
LArAthenaPool: replace LArDigitContainer_p2 by p3 as new default

Hello,

This MR is a follow-up of !56546 which introduced LArDigitContainer_p3 and the corresponding support in the LArDigitContainerCnv, but the p2->p3 change had not been made, to keep the output unchanged.
This MR deploys p3 as the new default.
It will be FT0V, and will require an update of q442 and q449 reference files.
I will let the CI fail, and whenenver possible, we can grab the CI output and bump up the WorkflowTestRunner versions.

Cheers,
Bertrand