FastCaloSim: changed the access to CaloDetDescrManager
Several clients in FastCaloSim used to either retrieve CaloDetDescrManager from DetStore,
or use the obsolete CaloDetDescrManager::instance(). Neither of these is compatible with
reading CaloDetDescrManager from Condition Store. However, given the complexity of the
clients, switching them to Condition Handles at this point would require a lot of
code refactoring.

Instead the following strategy has been implemented: the clients check at initialization
if the Detector Store contains the DD manager with the special key: "CaloMgrFCS". The
first client which tries to access the manager builds it, and records into DS with this key,
so that the other clients can use it. Given this is done in simulation, the alignments don't
change through the job. Hence the usage of Condition Handles is not really required.