AnalysisTauJets in CA version of DAOD_PHYSLITE
Draft: AnalysisTauJets in CA version of DAOD_PHYS

The relevant lines for including AnalysisTauJets in DAOD_PHYS are uncommented in the CA version. This works technically but it isnÂ't clear whether this configuration is safe, due to the use of public tools in the tau analysis sequence. Draft until this is clearer.