genCLIDDB Fix, 22.0 branch (2022.09.28.)
Stopped genCLIDDB from reading existing clid.db files.

By default, even though this was not the intention, the code was
always loading existing clid.db files together with the library
for which it was specifically meant to generate IDs for. This meant
that it was absolutely impossible to schema evolve xAOD types in
partial builds on top of a nightly.