Fixing over-writing of the extra content dictionary in CA version of PHYSLITE
Fixing over-writing of the extra content dictionary in CA version of PHYSLITE

This fixes a bug where the "AppendToDictionary" member of the slimming helper in PHYSLITE was over-written. This led to all of the trigger matching being omitted. With this fix the trigger matching containers are added to the CA version of PHYSLITE, bringing them into line with the legacy config. It should be noted that as things stand the trigger matching containers are _far_ too big (in both config versions), so this will need to be addressed in a separate MR.

P.S. The "AppendToDictionary" should probably be made into a method (as it name kind of implies already - but falsely) to avoid this sort of thing happening, but it will require many changes across dozens of python files that are already in flux due to the CA migration, so again, best to address separately.