Fix bug in pileup re-weighting auto-config code
Fix bug in pileup reweighting auto-config code

After switching to enums for storing campaign variable the enum
was used to build the pileup config filename instead of the enum
value. This caused the pileup tool to use a default config instead
of a dsid specific config. This MR switches to using the value of
the enum when building the pileup config.