TrigDecisionTool: compile regex for ChainGroup once
Store the compiled regex in the `ChainGroup` instead of re-compiling it
on each use. This also moves the `TrigDefs::Group` properties into the
constructor where they really belong.

Also improve the printout when updating the configuration to include
the total number of chain groups and the time required for the update.

Relates to ATR-23427.