Add option to skip decoding for NSW detectors
Add option to skip decoding for NSW detectors

Adding the option to skip the decoding of the NSW detectors after the container is created so that downstream algorithms still have the (empty) container available to satisfy the data dependencies.
This can help save some CPU time in the trigger when we aren't using the NSW hits in the reconstruction (ATR-26412).