TrigValTools: update lowmu input file
TrigValTools: update lowmu input file

Hello,

This MR is updating a low-mu RAW data file used in a trigger ART test.
The test is now crashing due to the recent update of default geometry and conditions used in the HLT (!56994, ATR-26453).
Moving to a Run3 data file solves the problem.

Cheers,
Bertrand