Fix potential memory leak in termination of the job.
Fix potential memory leak in tyermination of the job.

The moniotring class was allocating an analysis object which was used as normal,
but it was explicitlyt cleaning up at the end, and was leaving it to the default
garbage collection. This should heva been harmless, but this fix potentially corrects
that, although it may introduce other errors if initialise needs to be called again.

Potentially fixes part of https://its.cern.ch/jira/browse/ATR-26313