StgcRawDataMonitoring: implementation and improvement of occupancy plots per sTGC channel.
StgcRawDataMonitoring: implementation and improvement of occupancy plots per sTGC channel.

In this version of the package we implement occupancy plots for the different sTGC channels, i.e., pad, strip
and wire channels. In these plots we put togheter the ASide and CSide to reduce the number of output plots.
Also, in order to have a better distintion between sectors, we put a gap between them.

On the other hand, we performed several modifications to the code aiming a better readibility of it and also to reduce
the number of output plots.