Fixing missing variables in CA version of PHYS
Fixing an issue where some variables needed for data and MC were put in the MC-only cateogory in PHYS

In the migration of DAOD_PHYS to CA we did a clean up of variables that are only found in MC, to reduce the number of warnings. Unfortunately some variables that are also needed in data were accidentally put behind the isMC if statement, leading to missing variables from data. The variable comparison tests were done with MC so this was missed.

This is fixed in this MR.