TRTTransitionRadiation: support running with GeoModel initialized from SQLite database
TRTTransitionRadiation needs to build two G4 materials. In the original implementation
it used the StoredMaterialManager for building these G4 materials.

However, when GeoModel is initialized from the SQLite database, the StoredMaterialManager
is not available in the Detector Store. Instead, TRTTransitionRadiation now uses
a special table in SQLite database for building the G4 materials