Removing two obsolete lists in DAOD SlimmingHelper
Removing two obsolete lists in DAOD SlimmingHelper

The two lists of containers `FullListOfSmartContainers` and `StaticNamesAndTypes` are no longer needed in the slimming machinery, and are removed. The former was not used for anything, whereas the second was used in the event that the file metadata didn't have the list of named types. Since this should always be available, this can be safely dropped.