Workaround for CA-based ReSimulation jobs
Workaround for CA-based ReSimulation jobs

This ensures that the conditions run number is available during the
initialization again. The CA-based configuration behaviour is now the
same as the legacy configuration. Further debugging is required to see
if this is correct in the case of variable beam spot simulation.
Should fix ATLASSIM-6113.