Upgrading plugin capabilities in MadGraphControl
Upgrading plugin capabilities in MadGraphControl

MadGraph5_aMC@NLO allows the use of plugins that can be enabled by a
command-line flag. Much to my chagrin, some of them must be enabled on
the command line, and cannot be applied at run time. This enables new
functionality to apply a plugin of the user's choice as a part of the
command-line start up of mg5_amc.

At the same time, removing the old CVMFS plugin that we were
hard-coding. This should no longer be needed because of the more complex
shutils patch that we have gone for here.

Combined with a patch to MadGraph itself and a new deployment of
MadGraphModels, this should be all that's needed for plugins in master.
I'll go back and create an equivalent patch for 21.6 shortly.