introduce a dummy version of ShallowCopyDecorDeps for AnalysisBase (ATLASG-2405)
This is mostly to allow the jet reconstruction code to be run in
AnalysisBase.  It doesn't do anything, but is a placeholder to allow
dual-use code to use it in AthenaMT.  I didn't declare any properties,
on the assumption that they will not be needed (i.e. there will be no
attempt to set them from python configuration).