Allow the possibility to removed matched tracks from the reference selection
Allow the possibility to removed matched tracks from the reference selection

For eg the muonIso tracking, the point is to recostruct the tracks for isolation
not the primary muon, which has already been reconstructed and selected.

Therefore for the analysis it is useful to consider *only* the pereformance for
the non-muon tracks.

These changes add the possibility to specify a collection of eg muons, where the
reference tracks which match this tracks in this collection are removed from the
reference collection for the subsequent analysis