AthMonitorAlgorithm  - add a flag to be set if Alg is running in an the express stream
AthMonitorAlgorithm  - add a flag to be set if Alg is running in an the express stream

This is a preliminary commit. The express stream is added, but not yet used.

Adding an attribute to AthMonitorAlgorithm, bit not rebuilding the
derived Algorithms (which occur in many packages) causes a run time failure.

Once the new version of the AthMonitorAlgorithm is built, adding the use of
the attribute (and testing the new code) can be submiited in a future merge request.