Modification of V2.3 processing to work well with all clients
Upgraded V2.3 to V2.3.1

V2.3.1 has only change in location of version flag.
In V2.3 it is located in HanMetadata_/File_version.
Now it is located in HanMetadata_/File/Version_name.

This change is needed for HanMetadata.py to work fine, since
it expects from subdirectories of HanMetadata_ directory to have
two levels (not one).