GenericMonitoringTool: remove assert for path==None for online monitoring
Allow the usage of `path==None` also for online monitoring. This check
was only ever meant for the HLT to ensure developers specify an explicit
path (e.g. EXPERT). But for other online monitoring (e.g.
GlobalMonitoring) no such restriction should exist. Rather than
complicating the logic even further, remove this check. If someone were
to forget to specify the `path` in the HLT, the histogram would simply
not appear in the output.

Fixes ATR-26666 and ATLLONECAL-436.