AthExUnittest: Fix FPE test.
AthExAlgWithFPE is supposed to generate a FPE by division by zero.
But for this to work reliably, we must be sure the compiler cannot see
the division by zero at compile time.
Move the actual division to a separate compilation unit.

Fixes test failure in AthenaConfiguration in the clang build.