Change default setting of run3Commissioning flag for muon trigger
Change default setting of run3Commissioning flag for muon trigger

Setting the default value of the run3Commissioning flag for the new JO in the muon trigger to be False. This is to align with the same flag in the old-style JO which is False by default in the trigger. As discussed in ATR-26543, the different default behaviour was causing some confusion.