SUSYTools: fix IsRun3 behaviour
Hello,

As discussed in ATLSUSYBGF-645, there is a `IsRun3` property to configure SUSYTools for Run2/Run3.
This setting can be specified either via a property, or a config file.
Currently, the property is ignored as it's overriden by the config file.
With this MR, the user can now switch Run2/Run3 via the property, and use the same SUSYTools config file for both Run2/Run3.

The other change is related to a muon tool configuration.

Cheers,
Bertrand