Add merging script for EB weights xml files (ATR-26569)
Add merging script for EB weights xml files (ATR-26569)

This script merges the EB weights xml files.
Each xml file has a list of weights with ids and a list of events with assigned weight (by id).
Weights' ids are different between the files and the merging script unifies the ids of the weights.