Combining the thinning tools for PHYS and PHYSLITE into a single config
Draft: Combining the thinning tools for PHYS and PHYSLITE into a single config

ATLASG-2390

This MR combines the (currently separate) thinning requirements for PHYS and PHYSLITE into a common config script. This will help to keep them in synch, and simplifies the requirement that the config supports both AOD->PHYSLITE and PHYS->PHYSLITE workflows. It also reduces duplicated config code.